Refactor: Reorganize public API exports
Simplify module exports to expose submodules (helpers, errors, policy,
utils, backends, serialization, deserialization) rather than
individual items. Improve lxml import handling with cleaner warning.
