+++ src/hypomnema/xml/__init__.py
@@ -4,25 +4,6 @@ Provides namespace handling, encoding normalization, URI validation,
 and other XML-related utilities.
 """
 
-from .policy import (
-  ActionEnum,
-  Behavior,
-  DuplicateChildAction,
-  NamespacePolicy,
-  RaiseIgnore,
-  RaiseIgnoreDefault,
-  RaiseIgnoreDelete,
-  RaiseIgnoreForce,
-  RaiseIgnoreOverwrite,
-  RaiseNoneKeep,
-  XmlDeserializationPolicy,
-  XmlSerializationPolicy,
-)
-from .utils import (
-  QNameLike,
-  make_usable_path,
-  normalize_encoding,
-  validate_ncname,
-  validate_uri,
-  fast_validate_uri,
-)
+from . import policy, utils, backends, serialization, deserialization
+
+__all__ = ["policy", "utils", "backends", "serialization", "deserialization"]
