+++ tests/xml/backends/test_imports.py
@@ -1,37 +1,21 @@
-import subprocess
-import sys
-import textwrap
+import importlib
+from unittest.mock import patch
+
 import pytest
 
+import hypomnema.xml.backends
 
-@pytest.mark.skipif(
-  sys.platform == "win32", reason="Subprocess isolation test; shell behavior differs on Windows"
-)
-def test_lxml_missing_fallback():
-  """
-  Test that LxmlBackend is set to None and a warning is raised if lxml
-  cannot be imported.
 
-  Note: This runs in a subprocess to safely manipulate sys.modules.
-  Coverage is intentionally skipped for this test due to subprocess isolation.
-  """
-  code = textwrap.dedent("""
-        import sys
-        import warnings
+def test_lxml_unavailable_fallback():
+  lxml_mocks = {"lxml": None, "lxml.etree": None, "hypomnema.xml.backends.lxml": None}
 
-        # Simulate lxml absence
-        sys.modules['lxml'] = None
+  try:
+    with (
+      patch.dict("sys.modules", lxml_mocks),
+      pytest.warns(UserWarning, match="lxml not installed"),
+    ):
+      importlib.reload(hypomnema.xml.backends)
 
-        # Capture warnings
-        with warnings.catch_warnings(record=True) as w:
-            warnings.simplefilter("always")
-            import hypomnema.xml.backends as backends
-
-            assert len(w) == 1
-            assert "lxml not installed" in str(w[0].message)
-            assert backends.LxmlBackend is None
-    """)
-
-  result = subprocess.run([sys.executable, "-c", code], capture_output=True, text=True)
-
-  assert result.returncode == 0, f"Subprocess failed: {result.stderr}"
+    assert hypomnema.xml.backends.LxmlBackend is None
+  finally:
+    importlib.reload(hypomnema.xml.backends)
